fer-project/requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
efficientnet-pytorch==0.7.1

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
pillow>=10.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Data Augmentation
albumentations>=1.3.1
imgaug>=0.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Progress & Logging
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Face Detection
mtcnn>=0.1.1
facenet-pytorch>=2.5.3
dlib>=19.24.0

# Utilities
jupyter>=1.0.0
ipywidgets>=8.0.0